+++ src/common/enumerator.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import, annotations
+
+from enum import Enum
+
+
+class AnswerOrder(Enum):
+
+    NORMAL = "normal"
+    REVERSE = "reverse"