+++ src/common/rules.py
@@ -9,7 +9,9 @@ from typing import List, Dict, Any
 from wenet.model.user.common import Date
 from wenet.model.user.profile import WeNetUserProfile
 
+from common.enumerator import AnswerOrder
 from ws.models.survey import SurveyAnswer
+
 
 logger = logging.getLogger("wenet-survey-web-app.common.profile")
 
@@ -145,7 +147,7 @@ class CompetenceMeaningNumberRule(Rule):
                 if isinstance(self.question_code, str) and isinstance(self.category_name, str) and isinstance(self.variable_name, str)\
                         and isinstance(survey_answer.answers[self.question_code].answer, int):
                     answer_number = survey_answer.answers[self.question_code].answer
-                    answer_percent = (answer_number-1)/self.ceiling_value #line that transforms number into float percentage
+                    answer_percent = (answer_number-1)/(self.ceiling_value-1)  # line that transforms number into float percentage
                     value = None
                     if self.profile_attribute == "meanings":
                         value = {"name": self.variable_name, "category": self.category_name, "level": answer_percent}
@@ -241,6 +243,48 @@ class MaterialsFieldRule(Rule):
                         logger.warning(f"field type {type(survey_answer.answers[self.question_code].answer)} is not supported")
             else:
                 logger.debug(f"Trying to apply rule but question code [{self.question_code}] is not selected by user")
+        else:
+            logger.warning(f"Trying to apply rule but the user ID [{user_profile.profile_id}] does not match the user ID in the survey [{survey_answer.wenet_id}]")
+        return user_profile
+
+
+class CompetenceMeaningBuilderRule(Rule):
+
+    def __init__(self, order_mapping: Dict[str, AnswerOrder], variable_name: str, ceiling_value: int, category_name: str, profile_attribute: str):
+        self.order_mapping = order_mapping
+        self.variable_name = variable_name
+        self.ceiling_value = ceiling_value
+        self.category_name = category_name
+        self.profile_attribute = profile_attribute
+
+    def apply(self, user_profile: WeNetUserProfile, survey_answer: SurveyAnswer) -> WeNetUserProfile:
+        if self.check_wenet_id(user_profile, survey_answer):
+            all_answers_selected = all(question_code in survey_answer.answers for question_code in self.order_mapping.keys())
+            if all_answers_selected:
+                all_question_codes_str = all(isinstance(question_code, str) for question_code in self.order_mapping.keys())
+                all_answers_int = all(isinstance(survey_answer.answers[question_code].answer, int) for question_code in self.order_mapping.keys())
+                if all_question_codes_str and all_answers_int and isinstance(self.variable_name, str) \
+                        and isinstance(self.category_name, str) and isinstance(self.ceiling_value, int) and isinstance(self.profile_attribute, str):
+                    required_answers = []
+                    for question_code in self.order_mapping.keys():
+                        answer_number = survey_answer.answers[question_code].answer
+                        if self.order_mapping.get(question_code) == AnswerOrder.REVERSE:
+                            answer_number = (self.ceiling_value - answer_number) + 1
+                        required_answers.append(answer_number)
+                    # (A1+A2+A3+A4-4)/16 or (A1+A2+A3-3)/12 in case of equation changes, see this line
+                    number_value = (sum(required_answers)-len(required_answers))/(len(required_answers)*4)
+                    profile_entry = None
+                    if self.profile_attribute == "meanings":
+                        profile_entry = {"name": self.variable_name, "category": self.category_name, "level": number_value}
+                    elif self.profile_attribute == "competences":
+                        profile_entry = {"name": self.variable_name, "ontology": self.category_name, "level": number_value}
+                    else:
+                        logger.warning(f"{self.profile_attribute} field is not supported in the user profile")
+                    if profile_entry is not None:
+                        getattr(user_profile, self.profile_attribute).append(profile_entry)
+                        logger.debug(f"updated {self.profile_attribute} with {getattr(user_profile, self.profile_attribute)}")
+            else:
+                logger.debug(f"Not all necessary answers are selected to build {self.variable_name} attribute of the user {user_profile.profile_id}")
         else:
             logger.warning(f"Trying to apply rule but the user ID [{user_profile.profile_id}] does not match the user ID in the survey [{survey_answer.wenet_id}]")
         return user_profile